Refactor for presenter in user collect
