--- app/src/main/java/br/com/ecostage/mobilecollect/category/selection/Classification.kt
@@ -1,6 +1,0 @@
-package br.com.ecostage.mobilecollect.category.selection
-
-/**
- * Created by cmaia on 7/23/17.
- */
-data class Classification(val id : String?, val name: String, val parent: String?)
