--- app/src/main/java/br/com/ecostage/mobilecollect/category/selection/ClassificationView.kt
@@ -1,8 +1,0 @@
-package br.com.ecostage.mobilecollect.category.selection
-
-/**
- * Created by cmaia on 7/23/17.
- */
-interface ClassificationView {
-
-}
