+++ app/src/main/java/br/com/ecostage/mobilecollect/ui/profile/collect/UserCollectPresenter.kt
@@ -1,0 +1,8 @@
+package br.com.ecostage.mobilecollect.ui.profile.collect
+
+/**
+ * Created by cmaia on 7/23/17.
+ */
+interface UserCollectPresenter {
+    fun loadCollects()
+}
