+++ app/src/main/java/br/com/ecostage/mobilecollect/ui/profile/collect/UserCollectView.kt
@@ -1,0 +1,7 @@
+package br.com.ecostage.mobilecollect.ui.profile.collect
+
+/**
+ * Created by cmaia on 7/23/17.
+ */
+interface UserCollectView {
+}
